+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Refactor base logics
  - Use scriptable object for audio unit
  - Add more documentation comments
+ - Make pool holder as don't destroy on load
 
 # 0.1.0
  - Import base logics of this repository